raspberry-pi-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
pandas==2.1.4
aiofiles==23.2.1
python-multipart==0.0.6
email-validator==2.1.0
jinja2==3.1.2
aiosmtplib==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
redis==5.0.1

